apps/api/requirements.txt
# PricePin API — FastAPI + Mistral OCR + Supabase
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.27.0
supabase>=2.0.0
Pillow>=10.0.0
slowapi>=0.1.9
